+++ src/lexograph/layout/__init__.py
@@ -1,5 +1,15 @@
 """Layout: place ordered units in 2-D or 3-D space (walks, spirals, grids)."""
 
 from lexograph.layout.linear import linear_layout
+from lexograph.layout.spiral import spiral_layout, tangent_angles
+from lexograph.layout.walk import heading_angles, walk_layout
+from lexograph.layout.widths import rendered_widths
 
-__all__ = ["linear_layout"]
+__all__ = [
+    "linear_layout",
+    "walk_layout",
+    "heading_angles",
+    "spiral_layout",
+    "tangent_angles",
+    "rendered_widths",
+]